# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hajj Companion API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface RegisterInput {
  passportNo: string;
  fullNameAr: string;
  /** @nullable */
  fullNameEn?: string | null;
  /**
   * @minLength 2
   * @maxLength 2
   */
  nationality: string;
  phone: string;
  /** @minLength 6 */
  password: string;
  /** @nullable */
  tentZone?: string | null;
  /** @nullable */
  emergencyContact?: string | null;
}

export interface LoginInput {
  phone: string;
  password: string;
}

export interface RefreshInput {
  refreshToken: string;
}

export interface User {
  id: string;
  passportNo: string;
  fullNameAr: string;
  /** @nullable */
  fullNameEn?: string | null;
  nationality: string;
  phone: string;
  /** @nullable */
  tentZone?: string | null;
  /** @nullable */
  groupId?: string | null;
  /** @nullable */
  emergencyContact?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  accessToken: string;
  refreshToken: string;
  user: User;
}

export interface GroupInput {
  nameAr: string;
  /** @nullable */
  agency?: string | null;
  /** @nullable */
  maxMembers?: number | null;
}

export interface JoinGroupInput {
  inviteCode: string;
}

export interface Group {
  id: string;
  nameAr: string;
  /** @nullable */
  agency?: string | null;
  /** @nullable */
  leaderId?: string | null;
  inviteCode: string;
  maxMembers?: number;
  memberCount?: number;
  createdAt: string;
}

export interface GroupMember {
  id: string;
  fullNameAr: string;
  /** @nullable */
  fullNameEn?: string | null;
  phone: string;
  /** @nullable */
  tentZone?: string | null;
  /** @nullable */
  lastLat?: number | null;
  /** @nullable */
  lastLng?: number | null;
  /** @nullable */
  lastSeenAt?: string | null;
}

export interface LocationPoint {
  latitude: number;
  longitude: number;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  zone?: string | null;
  recordedAt: string;
  isOfflineSync?: boolean;
}

export interface LocationBatch {
  locations: LocationPoint[];
}

export interface SyncResult {
  synced: number;
  message?: string;
}

export interface MemberLocation {
  userId: string;
  fullNameAr: string;
  latitude: number;
  longitude: number;
  /** @nullable */
  zone?: string | null;
  recordedAt: string;
}

export type ManasikItemStatus =
  (typeof ManasikItemStatus)[keyof typeof ManasikItemStatus];

export const ManasikItemStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface ManasikItem {
  key: string;
  titleAr: string;
  titleEn: string;
  /** @nullable */
  descriptionAr?: string | null;
  order: number;
  /** @nullable */
  day?: number | null;
  status: ManasikItemStatus;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
}

export type ManasikProgressUpdateStatus =
  (typeof ManasikProgressUpdateStatus)[keyof typeof ManasikProgressUpdateStatus];

export const ManasikProgressUpdateStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface ManasikProgressUpdate {
  status: ManasikProgressUpdateStatus;
}

export interface PlaceInput {
  nameAr: string;
  /** @nullable */
  nameEn?: string | null;
  emoji?: string;
  latitude: number;
  longitude: number;
  /** @nullable */
  description?: string | null;
  color?: string;
  isShared?: boolean;
  /** @nullable */
  sharedWithGroup?: string | null;
}

export interface PlaceUpdate {
  /** @nullable */
  nameAr?: string | null;
  /** @nullable */
  nameEn?: string | null;
  /** @nullable */
  emoji?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  isShared?: boolean | null;
  /** @nullable */
  sharedWithGroup?: string | null;
}

export interface Place {
  id: string;
  userId: string;
  nameAr: string;
  /** @nullable */
  nameEn?: string | null;
  emoji: string;
  latitude: number;
  longitude: number;
  /** @nullable */
  description?: string | null;
  color: string;
  isShared: boolean;
  /** @nullable */
  sharedWithGroup?: string | null;
  visitCount: number;
  /** @nullable */
  lastVisitedAt?: string | null;
  /** @nullable */
  distanceFromUserM?: number | null;
  createdAt: string;
}

export type NavigationHistoryItemStatus =
  (typeof NavigationHistoryItemStatus)[keyof typeof NavigationHistoryItemStatus];

export const NavigationHistoryItemStatus = {
  active: "active",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface NavigationHistoryItem {
  id: string;
  destinationName: string;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  cancelledAt?: string | null;
  /** @nullable */
  distanceTotalM?: number | null;
  /** @nullable */
  distanceWalkedM?: number | null;
  status: NavigationHistoryItemStatus;
}

export interface PlaceDetail {
  id: string;
  userId: string;
  nameAr: string;
  /** @nullable */
  nameEn?: string | null;
  emoji: string;
  latitude: number;
  longitude: number;
  /** @nullable */
  description?: string | null;
  color: string;
  isShared: boolean;
  visitCount: number;
  /** @nullable */
  lastVisitedAt?: string | null;
  createdAt: string;
  navigationHistory: NavigationHistoryItem[];
}

export interface NavigationStart {
  /** @nullable */
  placeId?: string | null;
  destinationLat: number;
  destinationLng: number;
  destinationName: string;
}

export interface NavDestination {
  lat: number;
  lng: number;
  name: string;
  /** @nullable */
  emoji?: string | null;
}

export interface NavigationSession {
  sessionId: string;
  destination: NavDestination;
  distanceM: number;
  bearingDegrees: number;
  directionAr: string;
  etaMinutes: number;
}

export interface NavigationPosition {
  currentLat: number;
  currentLng: number;
}

export interface NavigationUpdate {
  distanceRemainingM: number;
  bearingDegrees: number;
  directionAr: string;
  etaMinutes: number;
  isArrived: boolean;
}

export interface NavigationComplete {
  walkedM: number;
  durationMinutes: number;
  messageAr: string;
}

export interface SosInput {
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  alertType?: string;
}

export type EmergencyAlertStatus =
  (typeof EmergencyAlertStatus)[keyof typeof EmergencyAlertStatus];

export const EmergencyAlertStatus = {
  active: "active",
  resolved: "resolved",
} as const;

export interface EmergencyAlert {
  id: string;
  userId: string;
  alertType: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  status: EmergencyAlertStatus;
  createdAt: string;
  /** @nullable */
  resolvedAt?: string | null;
}

/**
 * @nullable
 */
export type DashboardSummaryGroup = {
  id?: string;
  nameAr?: string;
  memberCount?: number;
  onlineCount?: number;
} | null;

/**
 * @nullable
 */
export type DashboardSummaryActiveEmergency = {
  id?: string;
  alertType?: string;
  status?: string;
} | null;

export interface ManasikSummary {
  total: number;
  completed: number;
  inProgress: number;
  pending: number;
  percentComplete: number;
}

export interface DashboardSummary {
  user: User;
  /** @nullable */
  group?: DashboardSummaryGroup;
  manasikSummary: ManasikSummary;
  nearbyPlaces: Place[];
  /** @nullable */
  activeEmergency?: DashboardSummaryActiveEmergency;
}

export type GetPlacesParams = {
  lat?: number;
  lng?: number;
};

export type GetDashboardParams = {
  lat?: number;
  lng?: number;
};