# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hajj Companion API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /places:
    get:
      operationId: getPlaces
      tags: [places]
      summary: List user's saved places (with optional distance sorting)
      security:
        - bearerAuth: []
      parameters:
        - name: lat
          in: query
          required: false
          schema:
            type: number
        - name: lng
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: Places list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Place"
    post:
      operationId: createPlace
      tags: [places]
      summary: Save a new personal place
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceInput"
      responses:
        "201":
          description: Place created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
  /places/group/{groupId}:
    get:
      operationId: getGroupPlaces
      tags: [places]
      summary: Get shared places from all group members
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Shared group places
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Place"
  /places/{id}:
    get:
      operationId: getPlace
      tags: [places]
      summary: Get place details with navigation history
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Place details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaceDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePlace
      tags: [places]
      summary: Update a saved place
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceUpdate"
      responses:
        "200":
          description: Updated place
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
    delete:
      operationId: deletePlace
      tags: [places]
      summary: Delete a saved place
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted

  # NAVIGATION
  /navigation/start:
    post:
      operationId: startNavigation
      tags: [navigation]
      summary: Start a navigation session to a place or coordinates
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NavigationStart"
      responses:
        "201":
          description: Navigation session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NavigationSession"
  /navigation/history:
    get:
      operationId: getNavigationHistory
      tags: [navigation]
      summary: Get user's past navigation sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Navigation history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NavigationHistoryItem"
  /navigation/{sessionId}/update:
    post:
      operationId: updateNavigation
      tags: [navigation]
      summary: Update current position and get navigation guidance
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NavigationPosition"
      responses:
        "200":
          description: Navigation update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NavigationUpdate"
  /navigation/{sessionId}/complete:
    post:
      operationId: completeNavigation
      tags: [navigation]
      summary: Mark navigation session as complete (arrived)
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Navigation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NavigationComplete"
  /navigation/{sessionId}/cancel:
    post:
      operationId: cancelNavigation
      tags: [navigation]
      summary: Cancel an active navigation session
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # EMERGENCY
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # AUTH
...
    PlaceInput:
      type: object
      required: [nameAr, latitude, longitude]
      properties:
        nameAr:
          type: string
        nameEn:
          type: ["string", "null"]
        emoji:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        description:
          type: ["string", "null"]
        color:
          type: string
        isShared:
          type: boolean
        sharedWithGroup:
          type: ["string", "null"]
          format: uuid
    PlaceUpdate:
      type: object
      properties:
        nameAr:
          type: ["string", "null"]
        nameEn:
          type: ["string", "null"]
        emoji:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        isShared:
          type: ["boolean", "null"]
        sharedWithGroup:
          type: ["string", "null"]
          format: uuid
    Place:
      type: object
      required: [id, userId, nameAr, emoji, latitude, longitude, color, isShared, visitCount, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        nameAr:
          type: string
        nameEn:
          type: ["string", "null"]
        emoji:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        description:
          type: ["string", "null"]
        color:
          type: string
        isShared:
          type: boolean
        sharedWithGroup:
          type: ["string", "null"]
          format: uuid
        visitCount:
          type: integer
        lastVisitedAt:
          type: ["string", "null"]
          format: date-time
        distanceFromUserM:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
    PlaceDetail:
      type: object
      required: [id, userId, nameAr, emoji, latitude, longitude, color, isShared, visitCount, createdAt, navigationHistory]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        nameAr:
          type: string
        nameEn:
          type: ["string", "null"]
        emoji:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        description:
          type: ["string", "null"]
        color:
          type: string
        isShared:
          type: boolean
        visitCount:
          type: integer
        lastVisitedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        navigationHistory:
          type: array
          items:
            $ref: "#/components/schemas/NavigationHistoryItem"

    # NAVIGATION
    NavigationStart:
      type: object
      required: [destinationLat, destinationLng, destinationName]
      properties:
        placeId:
          type: ["string", "null"]
          format: uuid
        destinationLat:
          type: number
        destinationLng:
          type: number
        destinationName:
          type: string
    NavigationSession:
      type: object
      required: [sessionId, destination, distanceM, bearingDegrees, directionAr, etaMinutes]
      properties:
        sessionId:
          type: string
          format: uuid
        destination:
          $ref: "#/components/schemas/NavDestination"
        distanceM:
          type: number
        bearingDegrees:
          type: number
        directionAr:
          type: string
        etaMinutes:
          type: integer
    NavDestination:
      type: object
      required: [lat, lng, name]
      properties:
        lat:
          type: number
        lng:
          type: number
        name:
          type: string
        emoji:
          type: ["string", "null"]
    NavigationPosition:
      type: object
      required: [currentLat, currentLng]
      properties:
        currentLat:
          type: number
        currentLng:
          type: number
    NavigationUpdate:
      type: object
      required: [distanceRemainingM, bearingDegrees, directionAr, etaMinutes, isArrived]
      properties:
        distanceRemainingM:
          type: number
        bearingDegrees:
          type: number
        directionAr:
          type: string
        etaMinutes:
          type: integer
        isArrived:
          type: boolean
    NavigationComplete:
      type: object
      required: [walkedM, durationMinutes, messageAr]
      properties:
        walkedM:
          type: number
        durationMinutes:
          type: number
        messageAr:
          type: string
    NavigationHistoryItem:
      type: object
      required: [id, destinationName, startedAt, status]
      properties:
        id:
          type: string
          format: uuid
        destinationName:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        cancelledAt:
          type: ["string", "null"]
          format: date-time
        distanceTotalM:
          type: ["number", "null"]
        distanceWalkedM:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, completed, cancelled]

    # EMERGENCY
...