# lib/api-zod/src/generated/types/navigationHistoryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hajj Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { NavigationHistoryItemStatus } from "./navigationHistoryItemStatus";

export interface NavigationHistoryItem {
  id: string;
  destinationName: string;
  startedAt: Date;
  /** @nullable */
  completedAt?: Date | null;
  /** @nullable */
  cancelledAt?: Date | null;
  /** @nullable */
  distanceTotalM?: number | null;
  /** @nullable */
  distanceWalkedM?: number | null;
  status: NavigationHistoryItemStatus;
}