# lib/api-zod/src/generated/types/navigationStart.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hajj Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface NavigationStart {
  /** @nullable */
  placeId?: string | null;
  destinationLat: number;
  destinationLng: number;
  destinationName: string;
}